backend/requirements.txt
Flask-Bcrypt==1.0.1
pymongo==4.10.1
loguru==0.7.3
python-dotenv==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
PyMuPDF>=1.24.0,<1.25.0
python-docx==1.1.2
textract-py3==2.1.1